move some modules to internal dir
Since kubebuilder's layout is now Standard Go Project Layout, some
modules have been re-located.　Since there is no specification as
to where the constants should be located, they were moved to the
top directory, similar to TopoLVM.
